# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "BenchmarkResult": {
        "properties": {
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "inference": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InferenceBenchmark"
              },
              {
                "type": "null"
              }
            ]
          },
          "inference_note": {
            "title": "Inference Note",
            "type": "string"
          },
          "memory_available_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Available Bytes"
          },
          "memory_copy_gbps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Single-thread memcpy bandwidth in GB/s (decimal gigabytes).",
            "title": "Memory Copy Gbps"
          },
          "memory_pressure_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Share of RAM in use at benchmark time (whole machine).",
            "title": "Memory Pressure Percent"
          },
          "memory_total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Total Bytes"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "ran_at": {
            "format": "date-time",
            "title": "Ran At",
            "type": "string"
          }
        },
        "required": [
          "ran_at",
          "duration_seconds",
          "memory_copy_gbps",
          "memory_total_bytes",
          "memory_available_bytes",
          "memory_pressure_percent",
          "inference",
          "inference_note",
          "notes"
        ],
        "title": "BenchmarkResult",
        "type": "object"
      },
...
      "CpuInfo": {
        "properties": {
          "architecture": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Architecture"
          },
          "logical_threads": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logical Threads"
          },
          "max_frequency_mhz": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Frequency Mhz"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "physical_cores": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Physical Cores"
          }
        },
        "required": [
          "model_name",
          "architecture",
          "physical_cores",
          "logical_threads",
          "max_frequency_mhz"
        ],
        "title": "CpuInfo",
        "type": "object"
      },
...
      "DownloadStatus": {
        "properties": {
          "bytes_done": {
            "title": "Bytes Done",
            "type": "integer"
          },
          "bytes_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bytes Total"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "model_id",
          "status",
          "bytes_done",
          "bytes_total",
          "error",
          "started_at",
          "finished_at"
        ],
        "title": "DownloadStatus",
        "type": "object"
      },
...
      "GpuInfo": {
        "properties": {
          "backend": {
            "$ref": "#/components/schemas/AcceleratorBackend",
            "default": "none"
          },
          "driver_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Driver Version"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "description": "Which probe produced this entry, e.g. 'nvidia-smi'.",
            "title": "Source",
            "type": "string"
          },
          "temperature_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature C"
          },
          "utilization_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Utilization Percent"
          },
          "vendor": {
            "$ref": "#/components/schemas/GpuVendor",
            "default": "unknown"
          },
          "vram_total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Total Bytes"
          },
          "vram_used_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vram Used Bytes"
          }
        },
        "required": [
          "index",
          "name",
          "vendor",
          "vram_total_bytes",
          "vram_used_bytes",
          "driver_version",
          "temperature_c",
          "utilization_percent",
          "backend",
          "source"
        ],
        "title": "GpuInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HardwareFit": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "recommended": {
            "title": "Recommended",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "recommended",
          "reasons"
        ],
        "title": "HardwareFit",
        "type": "object"
      },
      "HardwareReport": {
        "properties": {
          "benchmark": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BenchmarkResult"
              },
              {
                "type": "null"
              }
            ],
            "description": "The most recent measured benchmark, if one was run."
          },
          "cpu": {
            "$ref": "#/components/schemas/CpuInfo"
          },
          "detected_at": {
            "format": "date-time",
            "title": "Detected At",
            "type": "string"
          },
          "gpus": {
            "items": {
              "$ref": "#/components/schemas/GpuInfo"
            },
            "title": "Gpus",
            "type": "array"
          },
          "memory": {
            "$ref": "#/components/schemas/MemoryInfo"
          },
          "os": {
            "$ref": "#/components/schemas/OsInfo"
          },
          "tier": {
            "$ref": "#/components/schemas/TierEstimate"
          },
          "volumes": {
            "items": {
              "$ref": "#/components/schemas/StorageVolume"
            },
            "title": "Volumes",
            "type": "array"
          },
          "warnings": {
            "description": "Probes that failed or were unavailable. Shown to the user, not hidden.",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "detected_at",
          "os",
          "cpu",
          "memory",
          "gpus",
          "volumes",
          "tier",
          "benchmark",
          "warnings"
        ],
        "title": "HardwareReport",
        "type": "object"
      },
...
      "MemoryInfo": {
        "properties": {
          "available_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Bytes"
          },
          "total_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Bytes"
          }
        },
        "required": [
          "total_bytes",
          "available_bytes"
        ],
        "title": "MemoryInfo",
        "type": "object"
      },
...
      "ModelEntry": {
        "description": "A catalog model joined with local state, as the UI shows it.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "catalog": {
            "$ref": "#/components/schemas/CatalogModel"
          },
          "download": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DownloadStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "fit": {
            "$ref": "#/components/schemas/HardwareFit"
          },
          "installed": {
            "title": "Installed",
            "type": "boolean"
          },
          "license_accepted": {
            "title": "License Accepted",
            "type": "boolean"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "verified_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Sha256"
          }
        },
        "required": [
          "catalog",
          "installed",
          "active",
          "license_accepted",
          "file_path",
          "size_bytes",
          "verified_sha256",
          "fit",
          "download"
        ],
        "title": "ModelEntry",
        "type": "object"
      },
      "ModelIdBody": {
        "properties": {
          "model_id": {
            "title": "Model Id",
            "type": "string"
          }
        },
        "required": [
          "model_id"
        ],
        "title": "ModelIdBody",
        "type": "object"
      },
...
      "ModelsOverview": {
        "properties": {
          "active_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Model Id"
          },
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backend"
          },
          "loaded_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loaded Model Id"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "title": "Models",
            "type": "array"
          },
          "runtime_available": {
            "title": "Runtime Available",
            "type": "boolean"
          },
          "runtime_detail": {
            "title": "Runtime Detail",
            "type": "string"
          }
        },
        "required": [
          "runtime_available",
          "runtime_detail",
          "active_model_id",
          "loaded_model_id",
          "backend",
          "models"
        ],
        "title": "ModelsOverview",
        "type": "object"
      },
...
      "OsInfo": {
        "properties": {
          "machine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine"
          },
          "python_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Python Version"
          },
          "release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release"
          },
          "system": {
            "title": "System",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "system",
          "release",
          "version",
          "machine",
          "python_version"
        ],
        "title": "OsInfo",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MyAI Academy local service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/hardware": {
      "get": {
        "operationId": "read_hardware_api_hardware_get",
        "parameters": [
          {
            "description": "Re-run all probes instead of using the cache.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Re-run all probes instead of using the cache.",
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HardwareReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Hardware",
        "tags": [
          "hardware"
        ]
      }
    },
    "/api/hardware/benchmark": {
      "get": {
        "operationId": "read_benchmark_api_hardware_benchmark_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/BenchmarkResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Read Benchmark Api Hardware Benchmark Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Benchmark",
        "tags": [
          "hardware"
        ]
      },
      "post": {
        "description": "Run the short benchmark (spec \u00a730) and record the result.",
        "operationId": "run_hardware_benchmark_api_hardware_benchmark_post",
        "parameters": [
          {
            "description": "Also time the active local model, if one is installed.",
            "in": "query",
            "name": "include_inference",
            "required": false,
            "schema": {
              "default": true,
              "description": "Also time the active local model, if one is installed.",
              "title": "Include Inference",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Hardware Benchmark",
        "tags": [
          "hardware"
        ]
      }
    },
...
    "/api/models/active": {
      "post": {
        "operationId": "set_active_api_models_active_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelIdBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Active",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/downloads/{job_id}": {
      "get": {
        "operationId": "download_status_api_models_downloads__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Status",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/downloads/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_download_api_models_downloads__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Download",
        "tags": [
          "models"
        ]
      }
    },
...
    "/api/models/{model_id}": {
      "delete": {
        "operationId": "remove_model_api_models__model_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelsOverview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Model",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/{model_id}/accept-license": {
      "post": {
        "operationId": "accept_license_api_models__model_id__accept_license_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept License",
        "tags": [
          "models"
        ]
      }
    },
    "/api/models/{model_id}/download": {
      "post": {
        "operationId": "start_download_api_models__model_id__download_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Download",
        "tags": [
          "models"
        ]
      }
    },
...
  }
}